python/requirements.txt
# TAP A1 Talent Intake Agent Requirements

# Database
mysql-connector-python>=8.0.33

# LLM
google-generativeai>=0.3.0
google-genai>=1.0.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pypdf>=6.0.0
pdf2image>=1.16.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# DOC/DOCX Processing
python-docx>=1.0.0
docx2txt>=0.8
docx2pdf>=0.1.8
docxtpl>=0.16.7

# Web Scraping
playwright>=1.40.0

# Date Parsing
python-dateutil>=2.8.2
scikit-learn>=1.3.0

# Agent matching sémantique (A4 - embeddings)
# PyTorch 2.2+ pour register_pytree_node (transformers). Aligner torchvision/torchaudio.
torch>=2.2.0,<2.6.0
torchvision>=0.17.0
torchaudio>=2.2.0
# Combo stable Python 3.10 (évite l'erreur d'annotations numpy dans data_collator)
transformers==4.41.2
sentence-transformers==2.7.0
# Version compatible avec le combo ci-dessus
numpy==1.26.4

# Utilities
python-dotenv>=1.0.0
qrcode[pil]>=7.0

# API (optional)
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
httpx>=0.26.0
pydantic>=2.5.0

# Avatar TTS (B3 - page /avatar)
elevenlabs>=1.0.0

# Template Engine
Jinja2>=3.1.0

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Object Storage
minio>=7.2.0

# Production Server
gunicorn>=21.2.0

# Auth / Tokens
PyJWT>=2.8.0
